requirements.txt
# requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
groq==0.9.0
python-dotenv==1.0.1
gunicorn==22.0.0
firebase-admin==6.5.0
httpx==0.27.0
yfinance==0.2.40
feedparser==6.0.11